requirements.txt
streamlit
requests
matplotlib.pyplot
plotly.graph_objects
pandas
